Move Sensor class to more logical location
